chore: Move buildArgs to pkg/operator to be able to reuse it
